refactor(QueryParams): move getSafeQueryParams and addQueryParams to path to shared
