+++ src/server/lib/__tests__/queryParams.test.ts
@@ -1,15 +1,8 @@
 import {
   parseExpressQueryParams,
   addReturnUrlToPath,
-  getSafeQueryParams,
-  addQueryParamsToPath,
 } from '@/server/lib/queryParams';
 import { getConfiguration } from '@/server/lib/getConfiguration';
-import {
-  QueryParams,
-  SafeQueryParams,
-  UnsafeQueryParams,
-} from '@/shared/model/QueryParams';
 
 // mock configuration to return a default uri
 jest.mock('@/server/lib/getConfiguration', () => ({
@@ -113,124 +106,6 @@ describe('addReturnUrlToPath', () => {
       const expected =
         '/test/path?otherParam=b%3A%2F%2Ftest+b&returnUrl=a%3A%2F%2F%20test';
       expect(output).toEqual(expected);
-    });
-  });
-
-  describe('getSafeQueryParams', () => {
-    it('removes unsafe properties from the query object', () => {
-      const input: QueryParams = {
-        returnUrl: 'returnUrl',
-        clientId: 'clientId',
-        csrfError: true,
-        emailVerified: true,
-        encryptedEmail: 'encryptedEmail',
-        error: 'error',
-        ref: 'ref',
-        refViewId: 'refViewId',
-      };
-
-      const output = getSafeQueryParams(input);
-
-      const expected: SafeQueryParams = {
-        returnUrl: 'returnUrl',
-        clientId: 'clientId',
-        ref: 'ref',
-        refViewId: 'refViewId',
-      };
-
-      expect(output).toStrictEqual(expected);
-    });
-  });
-
-  describe('addQueryParamsToPath', () => {
-    it('adds safe query params to path without preexisting querystring', () => {
-      const input: QueryParams = {
-        returnUrl: 'returnUrl',
-        clientId: 'clientId',
-        csrfError: true,
-        emailVerified: true,
-        encryptedEmail: 'encryptedEmail',
-        error: 'error',
-        ref: 'ref',
-        refViewId: 'refViewId',
-      };
-
-      const output = addQueryParamsToPath('/test', input);
-
-      expect(output).toEqual(
-        '/test?clientId=clientId&ref=ref&refViewId=refViewId&returnUrl=returnUrl',
-      );
-    });
-
-    it('adds safe query params to path with preexisting querystring', () => {
-      const input: QueryParams = {
-        returnUrl: 'returnUrl',
-        clientId: 'clientId',
-        csrfError: true,
-        emailVerified: true,
-        encryptedEmail: 'encryptedEmail',
-        error: 'error',
-        ref: 'ref',
-        refViewId: 'refViewId',
-      };
-
-      const output = addQueryParamsToPath('/test?foo=bar', input);
-
-      expect(output).toEqual(
-        '/test?foo=bar&clientId=clientId&ref=ref&refViewId=refViewId&returnUrl=returnUrl',
-      );
-    });
-
-    it('adds safe query params to path without preexisting querystring, with manual unsafe values', () => {
-      const inputSafe: QueryParams = {
-        returnUrl: 'returnUrl',
-        clientId: 'clientId',
-        csrfError: true,
-        emailVerified: true,
-        encryptedEmail: 'encryptedEmail',
-        error: 'error',
-        ref: 'ref',
-        refViewId: 'refViewId',
-      };
-
-      const inputUnsafe: UnsafeQueryParams = {
-        csrfError: true,
-        encryptedEmail: 'an encrypted email',
-      };
-
-      const output = addQueryParamsToPath('/test', inputSafe, inputUnsafe);
-
-      expect(output).toEqual(
-        '/test?clientId=clientId&ref=ref&refViewId=refViewId&returnUrl=returnUrl&csrfError=true&encryptedEmail=an%20encrypted%20email',
-      );
-    });
-
-    it('adds safe query params to path with preexisting querystring, with manual unsafe values', () => {
-      const inputSafe: QueryParams = {
-        returnUrl: 'returnUrl',
-        clientId: 'clientId',
-        csrfError: true,
-        emailVerified: true,
-        encryptedEmail: 'encryptedEmail',
-        error: 'error',
-        ref: 'ref',
-        refViewId: 'refViewId',
-      };
-
-      const inputUnsafe: UnsafeQueryParams = {
-        csrfError: true,
-        encryptedEmail: 'an encrypted email',
-      };
-
-      const output = addQueryParamsToPath(
-        '/test?foo=bar',
-        inputSafe,
-        inputUnsafe,
-      );
-
-      expect(output).toEqual(
-        '/test?foo=bar&clientId=clientId&ref=ref&refViewId=refViewId&returnUrl=returnUrl&csrfError=true&encryptedEmail=an%20encrypted%20email',
-      );
     });
   });
 });
