+++ src/server/lib/queryParams.ts
@@ -1,12 +1,7 @@
-import {
-  QueryParams,
-  SafeQueryParams,
-  UnsafeQueryParams,
-} from '@/shared/model/QueryParams';
+import { QueryParams } from '@/shared/model/QueryParams';
 import { validateReturnUrl, validateRefUrl } from '@/server/lib/validateUrl';
 import { validateClientId } from '@/server/lib/validateClientId';
 import { FederationErrors } from '@/shared/model/Errors';
-import { stringify } from 'query-string';
 
 const validateEmailVerified = (emailVerified?: string): boolean | undefined => {
   if (!emailVerified) {
@@ -74,30 +69,4 @@ export const parseExpressQueryParams = (
 export const addReturnUrlToPath = (path: string, returnUrl: string): string => {
   const divider = path.includes('?') ? '&' : '?';
   return `${path}${divider}returnUrl=${encodeURIComponent(returnUrl)}`;
-};
-
-export const getSafeQueryParams = (params: QueryParams): SafeQueryParams => ({
-  returnUrl: params.returnUrl,
-  clientId: params.clientId,
-  ref: params.ref,
-  refViewId: params.refViewId,
-});
-
-export const addQueryParamsToPath = (
-  path: string,
-  params: QueryParams,
-  unsafeParams?: UnsafeQueryParams,
-): string => {
-  const divider = path.includes('?') ? '&' : '?';
-  const safeQueryString = stringify(getSafeQueryParams(params), {
-    skipNull: true,
-    skipEmptyString: true,
-  });
-  const unsafeQueryString = stringify(unsafeParams || {}, {
-    skipNull: true,
-    skipEmptyString: true,
-  });
-  return `${path}${divider}${safeQueryString}${
-    unsafeQueryString ? `&${unsafeQueryString}` : ''
-  }`;
 };
